move `getBlockByTime` to a util function
