+++ src/components/[guild]/collect/hooks/useNftDetails.ts
@@ -2,10 +2,10 @@ import { BigNumber } from "@ethersproject/bignumber"
 import { Contract } from "@ethersproject/contracts"
 import { JsonRpcBatchProvider } from "@ethersproject/providers"
 import { Chain, Chains, RPC } from "connectors"
-import { getBlockByTime } from "requirements/WalletActivity/hooks/useBlockNumberByTimestamp"
 import ERC721_ABI from "static/abis/erc721Abi.json"
 import useSWRImmutable from "swr/immutable"
 import base64ToObject from "utils/base64ToObject"
+import { getBlockByTime } from "utils/getBlockByTime"
 import { NULL_ADDRESS } from "utils/guildCheckout/constants"
 
 type NftStandard = "ERC-721" | "ERC-1155" | "Unknown"
